+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - New #63: Add nested mapping support via new `ObjectMap` class (@vjik)
 - Enh #99: Improve psalm annotation in `HydratorInterface::create()` method (@vjik)
+- Chg #103: Change PHP constraint in `composer.json` to `8.1 - 8.4` (@vjik)
 
 ## 1.5.0 September 17, 2024